@@ -2,8 +2,22 @@
   "name": "chrt-interpolations",
   "version": "0.0.9",
   "description": "Interpolation functions for Chrt",
-  "main": "dist/chrt-interpolations.js",
+  "keywords": [
+    "chrt",
+    "charts",
+    "interpolations",
+    "curves",
+    "visualization"
+  ],
+  "type": "module",
+  "main": "src/index.js",
   "module": "dist/chrt-interpolations.esm.js",
+  "jsdelivr": "dist/chrt-interpolations.min.js",
+  "unpkg": "dist/chrt-interpolations.min.js",
+  "exports": {
+    "umd": "./dist/chrt-interpolations.min.js",
+    "default": "./dist/chrt-interpolations.esm.js"
+  },
   "watch": {
     "build": {
       "patterns": [
@@ -17,7 +31,8 @@
     "postversion": "rollup -c",
     "build": "rollup -c",
     "lint": "eslint src --fix",
-    "watch": "npm-watch"
+    "watch": "npm-watch",
+    "prepare": "husky install"
   },
   "husky": {
     "hooks": {
@@ -36,18 +51,18 @@
   },
   "homepage": "chrt.io",
   "devDependencies": {
-    "@babel/core": "^7.12.10",
-    "@babel/plugin-transform-modules-commonjs": "^7.12.1",
-    "@babel/preset-env": "^7.12.11",
-    "@rollup/plugin-babel": "^5.2.2",
-    "@rollup/plugin-node-resolve": "^13.0.0",
-    "babel-jest": "^26.6.3",
+    "@babel/core": "^7.15.8",
+    "@babel/plugin-transform-modules-commonjs": "^7.15.4",
+    "@babel/preset-env": "^7.15.8",
+    "@rollup/plugin-babel": "^5.3.0",
+    "@rollup/plugin-node-resolve": "^13.0.5",
+    "babel-jest": "^27.2.5",
     "babel-plugin-root-import": "^6.6.0",
-    "chrt": "0.0.59",
-    "eslint": "^7.26.0",
-    "husky": "^4.2.5",
-    "jest": "^26.6.3",
-    "npm-watch": "^0.9.0",
+    "chrt": "0.0.107",
+    "eslint": "^8.0.0",
+    "husky": "^7.0.2",
+    "jest": "^27.2.5",
+    "npm-watch": "^0.11.0",
     "rollup": "2",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-node-resolve": "^5.2.0",
@@ -58,5 +73,6 @@
   "files": [
     "dist/**/*.js",
     "src/**/*.js"
-  ]
+  ],
+  "dependencies": {}
 }
